\name{PCA599}
\alias{PCA599}
\title{Principal Compenent Analysis}
\description{Calls code for \code{\line{PCA599} function that computs principal compnents for an input feature matrix.}
\usage{PCA599(data.matrix, 
    K = maximum.components)}

  \item{data.matrix}{Matirx of numeric data.}
  \item{K}{Maximum number of principal components to compute}


\value{List output from \code{\link{PCA599}} which represents a principal component dimension reduction
model.}

\author{Jared Benajmin Silverstein}


